resource-server/requirements/prod.txt
Django==3.0.3
djangorestframework==3.11.0
django-cors-headers==3.2.1
django-oauth-toolkit==1.2.0
gunicorn==20.0.4
gevent==1.4.0
psycopg2-binary==2.8.4
whitenoise==5.0.1
django-storages==1.9.1
dropbox==9.5.0